🧪test: Change API
Unfortunately the API the old API developer plan only worked on local servers, so I had to change it.

I spent a lot of time writing lines of code for each request that the new API delivers, since I have no way to control them. So I am finishing the commits for today.